requirements.txt
streamlit>=1.32.0
anthropic>=0.34.0
python-docx>=1.1.0
reportlab>=4.0.0
pdfplumber>=0.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-generativeai>=0.8.0
